update Java version

+++ src/test/java/com/upokecenter/util/XorShift128Plus.java
@@ -53,27 +53,27 @@ package com.upokecenter.util;
       synchronized (this.syncRoot) {
         while (length >= 8) {
           long nv = this.NextValue();
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           nv >>= 8;
-          bytes[offset++] = ((byte)nv);
+          bytes[offset++] = (byte)((byte)nv);
           length -= 8;
         }
         if (length != 0) {
           long nv = this.NextValue();
           while (length > 0) {
-            bytes[offset++] = ((byte)nv);
+            bytes[offset++] = (byte)((byte)nv);
             nv >>= 8;
             --length;
           }
